+++ src/main/arguments/arguments.py
@@ -40,13 +40,23 @@ class LlmArguments(CliArguments):
         self.max_parallel_requests = max_parallel_requests
 
     def validate(self) -> None:
-        dt_type_count = 0
-        dt_args = [self.use_4bit, self.use_8bit, self.is_bf16, self.is_fp16]
-        for dt_arg in dt_args:
-            if dt_arg:
-                dt_type_count += 1
-        if dt_type_count > 1:
-            raise ArgumentValidationException("only one of `use-4bit`, `use-8bit`, `is-bf16` or `is-fp16` data type options can be enabled at any instance in time")
+        # Quantization options (4bit/8bit) are mutually exclusive with each other
+        quantization_count = 0
+        quantization_args = [self.use_4bit, self.use_8bit]
+        for quant_arg in quantization_args:
+            if quant_arg:
+                quantization_count += 1
+        if quantization_count > 1:
+            raise ArgumentValidationException("only one of `use-4bit` or `use-8bit` can be enabled at any instance in time")
+
+        # Precision options (fp16/bf16) are mutually exclusive with each other
+        precision_count = 0
+        precision_args = [self.is_bf16, self.is_fp16]
+        for prec_arg in precision_args:
+            if prec_arg:
+                precision_count += 1
+        if precision_count > 1:
+            raise ArgumentValidationException("only one of `is-bf16` or `is-fp16` can be enabled at any instance in time")
 
         if self.padding_side is not None and not (self.padding_side == 'right' or self.padding_side == 'left'):
             raise ArgumentValidationException("`padding-side` must be one of either 'None', 'left' or 'right'")
@@ -87,13 +97,23 @@ class TunerFunctionArguments(CliArguments):
         if not is_valid:
             raise ArgumentValidationException("'Tuner Arguments' are missing required properties")
 
-        dt_type_count = 0
-        dt_args = [self.use_4bit, self.use_8bit, self.is_bf16, self.is_fp16]
-        for dt_arg in dt_args:
-            if dt_arg:
-                dt_type_count += 1
-        if dt_type_count > 1:
-            raise ArgumentValidationException("only one of `use-4bit`, `use-8bit`, `is-bf16` or `is-fp16` data type options can be enabled at any instance in time")
+        # Quantization options (4bit/8bit) are mutually exclusive with each other
+        quantization_count = 0
+        quantization_args = [self.use_4bit, self.use_8bit]
+        for quant_arg in quantization_args:
+            if quant_arg:
+                quantization_count += 1
+        if quantization_count > 1:
+            raise ArgumentValidationException("only one of `use-4bit` or `use-8bit` can be enabled at any instance in time")
+
+        # Precision options (fp16/bf16) are mutually exclusive with each other
+        precision_count = 0
+        precision_args = [self.is_bf16, self.is_fp16]
+        for prec_arg in precision_args:
+            if prec_arg:
+                precision_count += 1
+        if precision_count > 1:
+            raise ArgumentValidationException("only one of `is-bf16` or `is-fp16` can be enabled at any instance in time")
 
         if self.padding_side is not None and not (self.padding_side == 'right' or self.padding_side == 'left'):
             raise ArgumentValidationException("`padding-side` must be one of either 'None', 'left' or 'right'")